feat: 코스 관련 유틸리티 함수 경로 수정 및 불필요한 코드 삭제
